+++ vike/client/client-routing-runtime/history.ts
@@ -6,7 +6,8 @@ export { monkeyPatchHistoryAPI }
 export type { HistoryInfo }
 export type { ScrollPosition }
 
-import { assert, assertUsage, getCurrentUrl, getGlobalObject, hasProp, isObject } from './utils.js'
+import { getCurrentUrl } from '../shared/getCurrentUrl.js'
+import { assert, assertUsage, getGlobalObject, hasProp, isObject } from './utils.js'
 
 initHistoryState() // we redundantly call initHistoryState() to ensure it's called early
 const globalObject = getGlobalObject('history.ts', { previous: getHistoryInfo() })
@@ -68,7 +69,7 @@ function getState(): StateEnhanced {
   // - Therefore, we have to monkey patch history.pushState() and history.replaceState()
   // - Therefore, we need the assert() below to ensure history.state has been enhanced by Vike
   //   - If users stumble upon this assert() then let's make it a assertUsage()
-  assert(isVikeEnhanced(state))
+  assert(isVikeEnhanced(state), { state })
   return state
 }
 function getStateNotEnhanced(): StateNotEnhanced {