+++ vike/client/client-routing-runtime/skipLink.ts
@@ -1,18 +1,20 @@
 export { skipLink }
+export { isSameAsCurrentUrl }
 
+import { normalizeClientSideUrl } from '../shared/normalizeClientSideUrl.js'
 import { getBaseServer } from './getBaseServer.js'
 import { assert, parseUrl, isBaseServer, isUrl, isUrlExternal } from './utils.js'
 
 function skipLink(linkTag: HTMLElement): boolean {
-  const url = linkTag.getAttribute('href')
+  const href = linkTag.getAttribute('href')
   return (
-    url === null ||
-    !isUrl(url) ||
-    url === '' ||
-    isUrlExternal(url) ||
-    isHashUrl(url) ||
+    href === null ||
+    !isUrl(href) ||
+    href === '' ||
+    isUrlExternal(href) ||
+    isSamePageHashLink(href) ||
     isNewTabLink(linkTag) ||
-    !hasBaseServer(url) ||
+    !hasBaseServer(href) ||
     // Purposely last because disableAutomaticLinkInterception will be removed in the next major release
     !isVikeLink(linkTag)
   )
@@ -34,20 +36,24 @@ function isNewTabLink(linkTag: HTMLElement) {
   const rel = linkTag.getAttribute('rel')
   return target === '_blank' || target === '_external' || rel === 'external' || linkTag.hasAttribute('download')
 }
-function isHashUrl(url: string) {
-  if (url.startsWith('#')) {
+function isSamePageHashLink(href: string) {
+  if (
+    href.includes('#') &&
+    normalizeClientSideUrl(href, { withoutHash: true }) ===
+      normalizeClientSideUrl(window.location.href, { withoutHash: true })
+  ) {
     return true
   }
-  const removeHash = (url: string) => url.split('#')[0]
-  if (url.includes('#') && removeHash(url) === removeHash(window.location.pathname)) {
-    return true
-  }
+  assert(!href.startsWith('#'))
   return false
 }
-function hasBaseServer(url: string): boolean {
+function isSameAsCurrentUrl(href: string) {
+  return normalizeClientSideUrl(href) === normalizeClientSideUrl(window.location.href)
+}
+function hasBaseServer(href: string): boolean {
   const baseServer = getBaseServer()
   assert(isBaseServer(baseServer))
-  const { hasBaseServer } = parseUrl(url, baseServer)
+  const { hasBaseServer } = parseUrl(href, baseServer)
   return hasBaseServer
 }