+++ vike/utils/parseUrl.ts
@@ -42,11 +42,11 @@ type UrlPublic = {
   /** The URL search parameters array, e.g. `{ fruit: ['apple', 'orange'] }` for `https://example.com?fruit=apple&fruit=orange` **/
   searchAll: Record<string, string[]>
   /** The URL search parameterer string, e.g. `?details=yes` in `https://example.com/product/42?details=yes#reviews` */
-  searchOriginal: null | string
+  searchOriginal: null | `?${string}`
   /** The URL hash, e.g. `reviews` in `https://example.com/product/42?details=yes#reviews` */
   hash: string
   /** The URL hash string, e.g. `#reviews` in `https://example.com/product/42?details=yes#reviews` */
-  hashOriginal: null | string
+  hashOriginal: null | `#${string}`
 
   // TODO/v1-release: remove
   /** @deprecated */
@@ -61,16 +61,16 @@ function parseUrl(url: string, baseServer: string): UrlPrivate {
   assert(baseServer.startsWith('/'))
 
   // Hash
-  const [urlWithoutHash, ...hashList] = url.split('#')
+  const [urlWithoutHash, ...h] = url.split('#')
   assert(urlWithoutHash !== undefined)
-  const hashOriginal = ['', ...hashList].join('#') || null
+  const hashOriginal = (['', ...h].join('#') as undefined | `#${string}`) || null
   assert(hashOriginal === null || hashOriginal.startsWith('#'))
   const hash = hashOriginal === null ? '' : decodeSafe(hashOriginal.slice(1))
 
   // Search
   const [urlWithoutHashNorSearch, ...searchList] = urlWithoutHash.split('?')
   assert(urlWithoutHashNorSearch !== undefined)
-  const searchOriginal = ['', ...searchList].join('?') || null
+  const searchOriginal = (['', ...searchList].join('?') as undefined | `?${string}`) || null
   assert(searchOriginal === null || searchOriginal.startsWith('?'))
   const search: Record<string, string> = {}
   const searchAll: Record<string, string[]> = {}
@@ -159,14 +159,14 @@ function getPathnameAbsoluteWithBase(
 
     let base: string
     if (baseURI) {
-      const baseURIPathaname = parseOrigin(baseURI.split('?')[0]!).pathname
+      const baseURIPathaname = parseOrigin(baseURI.split('?')[0]!.split('#')[0]!).pathname
       base = baseURIPathaname
     } else {
       base = baseServer
     }
 
     const pathnameAbsoluteWithBase = resolveUrlPathnameRelative(url, base)
-    return { protocol: null, origin: null, pathnameAbsoluteWithBase: pathnameAbsoluteWithBase }
+    return { protocol: null, origin: null, pathnameAbsoluteWithBase }
   }
 }
 function parseOrigin(url: string): { pathname: string; origin: null | string; protocol: null | string } {